rbtree: implement Red-Black Tree algorithms
This Red-Black Tree (RBT) library is implemented as linked list are in the
Linux kernel. In order to use the RBT to store your data, each data
element contains one  `struct rbtree_node` which will be used to link
your data element in the RBT.

This library is not (yet) optimised.

Signed-off-by: Laurent Fazio <laurent.fazio@dev-at-home.fr>